refactor(common): extract reusable LogoImage component and fix edit profile logo rendering
